pyproject.toml
[tool.poetry]
name = "trading-platform"
version = "0.1.0"
description = ""
authors = ["StasHunter92 <stashunter92@gmail.com>"]
readme = "README.md"
packages = [{include = "trading_platform"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"
psycopg2-binary = "^2.9.6"
django-filter = "^23.2"
gunicorn = "^20.1.0"
drf-spectacular = "^0.26.2"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
